Move the OutboxProcessor to the production package
