+++ Libraries/CO.CDP.MQ.Tests/Outbox/OutboxProcessorTest.cs
@@ -66,19 +66,3 @@ public class OutboxProcessorTest
         UpdatedOn = DateTimeOffset.UtcNow
     };
 }
-
-class OutboxProcessor(IPublisher publisher, IOutboxMessageRepository outbox)
-{
-    public async Task Execute(int count)
-    {
-        var messages = await outbox.FindOldest(count);
-        messages.ForEach(PublishMessages);
-    }
-
-    private async void PublishMessages(OutboxMessage m)
-    {
-        await publisher.Publish(m);
-        m.Published = true;
-        await outbox.SaveAsync(m);
-    }
-}
