+++ Libraries/CO.CDP.MQ/Outbox/OutboxProcessor.cs
@@ -1,0 +1,17 @@
+namespace CO.CDP.MQ.Outbox;
+
+public class OutboxProcessor(IPublisher publisher, IOutboxMessageRepository outbox)
+{
+    public async Task Execute(int count)
+    {
+        var messages = await outbox.FindOldest(count);
+        messages.ForEach(PublishMessages);
+    }
+
+    private async void PublishMessages(OutboxMessage m)
+    {
+        await publisher.Publish(m);
+        m.Published = true;
+        await outbox.SaveAsync(m);
+    }
+}
